TS conversion starting point (#783)
