+++ extension/data/background/index.ts
