+++ extension/data/modules/betterbuttons.js
@@ -1,7 +1,7 @@
 import $ from 'jquery';
 
 import {Module} from '../tbmodule.js';
-import * as TBApi from '../tbapi.js';
+import * as TBApi from '../tbapi.ts';
 import * as TBCore from '../tbcore.js';
 import {getSettingAsync} from '../tbstorage.js';
