+++ extension/data/modules/modbar.js
@@ -5,7 +5,7 @@ import * as TBStorage from '../tbstorage.js';
 import * as TBui from '../tbui.js';
 import * as TBHelpers from '../tbhelpers.js';
 import * as TBCore from '../tbcore.js';
-import * as TBApi from '../tbapi.js';
+import * as TBApi from '../tbapi.ts';
 
 // Hold onto the modbarExists resolver so we can call it when the time is right
 let resolveModbarExists = null;
