+++ extension/data/modules/modmatrix.js
@@ -3,7 +3,7 @@ import SnuOwnd from 'snuownd';
 
 import {Module} from '../tbmodule.js';
 import * as TBStorage from '../tbstorage.js';
-import * as TBApi from '../tbapi.js';
+import * as TBApi from '../tbapi.ts';
 import * as TBCore from '../tbcore.js';
 import * as TBui from '../tbui.js';
