+++ extension/data/modules/notifier.js
@@ -3,7 +3,7 @@ import $ from 'jquery';
 
 import {Module} from '../tbmodule.js';
 import * as TBStorage from '../tbstorage.js';
-import * as TBApi from '../tbapi.js';
+import * as TBApi from '../tbapi.ts';
 import * as TBHelpers from '../tbhelpers.js';
 import * as TBCore from '../tbcore.js';
