+++ extension/data/modules/removalreasons.js
@@ -2,7 +2,7 @@ import $ from 'jquery';
 import SnuOwnd from 'snuownd';
 
 import {Module} from '../tbmodule.js';
-import * as TBApi from '../tbapi.js';
+import * as TBApi from '../tbapi.ts';
 import * as TBHelpers from '../tbhelpers.js';
 import * as TBCore from '../tbcore.js';
 import * as TBStorage from '../tbstorage.js';
