+++ extension/data/modules/trouble.js
@@ -2,7 +2,7 @@ import $ from 'jquery';
 
 import {Module} from '../tbmodule.js';
 import * as TBCore from '../tbcore.js';
-import {getCurrentUser} from '../tbapi.js';
+import {getCurrentUser} from '../tbapi.ts';
 
 const self = new Module({
     name: 'Trouble Shooter',
