+++ package.json
@@ -29,6 +29,10 @@
     "devDependencies": {
         "@rollup/plugin-commonjs": "^25.0.3",
         "@rollup/plugin-node-resolve": "^15.1.0",
+        "@rollup/plugin-typescript": "^11.1.2",
+        "@tsconfig/recommended": "^1.0.2",
+        "@types/jquery": "^3.5.16",
+        "@types/webextension-polyfill": "^0.10.1",
         "cross-env": "^7.0.3",
         "del-cli": "^5.0.0",
         "docdash": "^2.0.1",
@@ -37,6 +41,7 @@
         "inquirer": "^9.2.9",
         "jsdoc": "^4.0.2",
         "rollup": "^3.2.3",
-        "rollup-plugin-copy": "^3.4.0"
+        "rollup-plugin-copy": "^3.4.0",
+        "typescript": "^5.1.6"
     }
 }
